pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobsuche-py"
version = "0.1.0"
description = "Job search automation tool"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
    "types-pyyaml>=6.0.0",
    "pytest>=8.0.0",
    "pre-commit>=3.5.0",
    "coverage>=7.0.0",
]

[tool.ruff]
# Set the maximum line length
line-length = 100

# Assume Python 3.10+
target-version = "py310"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    "htmlcov",
]

[tool.ruff.lint]
# Enable rule sets:
# E/W = pycodestyle errors/warnings
# F = pyflakes
# I = isort (import sorting)
# N = pep8-naming
# UP = pyupgrade (modernize Python code)
# B = flake8-bugbear (find likely bugs)
# C4 = flake8-comprehensions
# SIM = flake8-simplify
# RUF = Ruff-specific rules
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]

# Ignore specific rules if needed
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]
# Allow print statements in scripts
"main.py" = ["T201"]
"tests/**/*.py" = ["S101"]  # Allow assert statements in tests

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true when ready to enforce
disallow_incomplete_defs = false  # Set to true when ready to enforce
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_optional = true

# Start with lenient settings, gradually tighten
# disallow_any_generics = true
# disallow_subclassing_any = true

[[tool.mypy.overrides]]
# Ignore missing imports for third-party packages without stubs
module = [
    "bs4.*",
    "yaml.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (make real network calls, require API key)",
]
# Default: skip integration tests unless explicitly requested
addopts = "-m 'not integration'"

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    ".venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
